fix(refactor): utils moved under lib
