+++ packages/react-components/src/lib/ol-map/source/stories/wmts.stories.tsx
@@ -3,7 +3,7 @@ import { Map } from '../../map';
 import { getWMTSOptions, TileWMTS } from '../wmts';
 import { TileOsm } from '../osm';
 import { TileLayer } from '../../layers/tile-layer';
-import { CSFStory } from '../../../../utils/story';
+import { CSFStory } from '../../../utils/story';
 
 
 const wmtsOptions = getWMTSOptions({
