+++ packages/react-components/src/lib/ol-map/source/stories/xyz.stories.tsx
@@ -3,7 +3,7 @@ import { Map } from '../../map';
 import { getXYZOptions, TileXYZ } from '../xyz';
 import { TileOsm } from '../osm';
 import { TileLayer } from '../../layers/tile-layer';
-import { CSFStory } from '../../../../utils/story';
+import { CSFStory } from '../../../utils/story';
 
 const xyzOptions = getXYZOptions({
   url:
